[🔨 Refactor] MainView NavigationPath 삭제
